Split LtsTimeStepper into its own header
The circular dependency that required it to be in TimeStepper.hpp is
gone with the switch to new-style factory for TimeStepper.
