+++ src/Evolution/Initialization/Evolution.hpp
@@ -24,6 +24,7 @@
 #include "Time/Tags.hpp"
 #include "Time/Time.hpp"
 #include "Time/TimeStepId.hpp"
+#include "Time/TimeSteppers/LtsTimeStepper.hpp"
 #include "Time/TimeSteppers/TimeStepper.hpp"
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
