+++ src/Time/TimeSteppers/CMakeLists.txt
@@ -25,6 +25,7 @@ spectre_target_headers(
   Cerk5.hpp
   DormandPrince5.hpp
   Factory.hpp
+  LtsTimeStepper.hpp
   RungeKutta3.hpp
   RungeKutta4.hpp
   TimeStepper.hpp
